+++ R/noteinfo.R
@@ -110,10 +110,9 @@ info_annotation <- function(x){
 #' The class is often used by other functions, and functions that output a
 #' note info string attach the \code{noteinfo} class.
 #'
-#' For \code{as_noteinfo} and in general for functions that accept
-#' the \code{format} argument, when it is \code{NULL}, the timestep delimiter
-#' format is inferred from the note info string input. When unclear, such as
-#' with phrase objects, the default is space-delimited time.
+#' When \code{format = NULL}, the timestep delimiter format is inferred
+#' from the note info string input. When unclear, such as with phrase objects,
+#' the default is space-delimited time.
 #'
 #' @param x character, a note info string.
 #' @param format \code{NULL} or character, the timestep delimiter format,